+++ packages/ui/app/src/docs/VersionDropdown.tsx
@@ -1,7 +1,8 @@
 import { type DocsNode } from "@fern-api/fdr-sdk";
 import { getVersionAvailabilityLabel } from "@fern-ui/app-utils";
-import classNames from "classnames";
-import { FernMenu, FernMenuItem } from "../components/FernMenu";
+import { CaretDownIcon } from "@radix-ui/react-icons";
+import { FernButton } from "../components/FernButton";
+import { FernDropdown } from "../components/FernDropdown";
 import { useNavigationContext } from "../navigation-context";
 
 export declare namespace VersionDropdown {
@@ -25,40 +26,24 @@ export const VersionDropdown: React.FC<VersionDropdown.Props> = ({
     const { basePath } = useNavigationContext();
     return (
         <div className="flex w-32">
-            <FernMenu text={selectedVersionName ?? ""}>
-                {versions.map(({ info: { id: versionName, availability, slug } }) => (
-                    <FernMenuItem
-                        key={versionName}
-                        selected={slug === selectedVersionSlug}
-                        href={createSlugHref(basePath, slug)}
-                    >
-                        {(active) => (
-                            <>
-                                <span className="font-mono text-sm font-normal">{versionName}</span>
-                                {availability != null && (
-                                    <span
-                                        className={classNames(
-                                            "rounded px-1 py-0.5 text-[11px] font-normal border",
-                                            {
-                                                "bg-accent-highlight border-transparent":
-                                                    slug === selectedVersionSlug && !active,
-                                                "bg-tag-default border-transparent":
-                                                    slug !== selectedVersionSlug && !active,
-                                            },
-                                            {
-                                                "border-accent-primary-light/75 dark:border-accent-primary-dark/75":
-                                                    active,
-                                            },
-                                        )}
-                                    >
-                                        {getVersionAvailabilityLabel(availability)}
-                                    </span>
-                                )}
-                            </>
-                        )}
-                    </FernMenuItem>
-                ))}
-            </FernMenu>
+            <FernDropdown
+                value={selectedVersionSlug}
+                options={versions.map(({ info: { id: versionName, availability, slug } }) => ({
+                    type: "value",
+                    label: versionName,
+                    helperText: availability != null ? getVersionAvailabilityLabel(availability) : undefined,
+                    value: slug,
+                    disabled: availability == null,
+                    href: createSlugHref(basePath, slug),
+                }))}
+            >
+                <FernButton
+                    intent="primary"
+                    variant="outlined"
+                    text={selectedVersionName}
+                    rightIcon={<CaretDownIcon className="transition-transform data-[state=open]:rotate-180" />}
+                />
+            </FernDropdown>
         </div>
     );
 };